refactor: move exports to index.ts
