+++ src/features/course-main/utils/index.ts
@@ -1,0 +1,2 @@
+export { hasDayInDate } from './has-day';
+export { setLabel } from './set-label';
